+++ composeApp/src/commonMain/kotlin/com/jermey/navplayground/demo/DI.kt
@@ -1,7 +1,7 @@
 package com.jermey.navplayground.demo
 
 import com.jermey.feature1.resultdemo.Feature1Module
-import com.jermey.navplayground.demo.app.sample.showcase.destinations.veeeeery.looong.packages.names.length.test.destinations.MainTabs
+import com.jermey.navplayground.navigation.MainTabs
 import com.jermey.quo.vadis.annotations.NavigationRoot
 import com.jermey.quo.vadis.core.navigation.config.NavigationConfig
 import com.jermey.quo.vadis.core.navigation.config.navigationConfig